refactor to `internal`
