+++ packages/effect/src/internal/stateMachineErrors.ts
@@ -1,0 +1,38 @@
+import type * as Cause from "../Cause.ts"
+import * as Data from "../Data.ts"
+
+/**
+ * Error returned when an event has no handler for the current state.
+ *
+ * @category errors
+ * @since 4.0.0
+ */
+export class UnhandledEventError extends Data.TaggedError("UnhandledEventError")<{
+  readonly machineId: string | undefined
+  readonly state: string
+  readonly event: string
+}> {}
+
+/**
+ * Error returned when a state machine does not stabilize within the maximum
+ * number of macrostep iterations.
+ *
+ * @category errors
+ * @since 4.0.0
+ */
+export class InfiniteTransitionError extends Data.TaggedError("InfiniteTransitionError")<{
+  readonly machineId: string | undefined
+  readonly state: string
+  readonly maxIterations: number
+}> {}
+
+/**
+ * Error returned when a state machine fails while running startup lifecycle
+ * logic after the initial state has been computed.
+ *
+ * @category errors
+ * @since 4.0.0
+ */
+export class StartupError extends Data.TaggedError("StartupError")<{
+  readonly cause: Cause.Cause<unknown>
+}> {}
