+++ packages/effect/src/unstable/actors/StateMachine.ts
@@ -4,21 +4,18 @@
  * @since 4.0.0
  */
 
-import type * as Cause from "../../Cause.ts"
-import * as Context from "../../Context.ts"
-import * as Data from "../../Data.ts"
 import * as Effect from "../../Effect.ts"
-import * as Exit from "../../Exit.ts"
-import * as HashMap from "../../HashMap.ts"
 import { PipeInspectableProto } from "../../internal/core.ts"
-import * as Option from "../../Option.ts"
+import type { InfiniteTransitionError, StartupError, UnhandledEventError } from "../../internal/stateMachineErrors.ts"
 import type { Pipeable } from "../../Pipeable.ts"
 import { hasProperty } from "../../Predicate.ts"
-import * as Ref from "../../Ref.ts"
-import * as Schema from "../../Schema.ts"
-import * as Scope from "../../Scope.ts"
-import * as Stream from "../../Stream.ts"
+import type * as Schema from "../../Schema.ts"
+import type * as Scope from "../../Scope.ts"
 import * as ActorModule from "./Actor.ts"
+import * as internalActor from "./internal/stateMachineActor.ts"
+import * as Model from "./internal/stateMachineModel.ts"
+import { ActorRuntime } from "./internal/stateMachineRuntime.ts"
+import * as internalRuntime from "./internal/stateMachineRuntime.ts"
 
 /**
  * String literal type used as the runtime type identifier for `StateMachine`
@@ -36,9 +33,6 @@ export type TypeId = "~effect/StateMachine"
  * @since 4.0.0
  */
 export const TypeId: TypeId = "~effect/StateMachine"
-
-const TargetTypeId = "~effect/StateMachine/Target"
-const CompletedOutputsTypeId: unique symbol = Symbol("effect/StateMachine/CompletedOutputs")
 
 type IsAny<A> = 0 extends (1 & A) ? true : false
 
@@ -90,69 +84,42 @@ export interface Machine<
   readonly initial: (...args: [...Machine.InputArgs<Input>]) => Machine.InitialResult<States, InitialE, InitialR>
 }
 
-/**
- * Error returned when an event has no handler for the current state.
- *
- * @category errors
- * @since 4.0.0
- */
-export class UnhandledEventError extends Data.TaggedError("UnhandledEventError")<{
-  readonly machineId: string | undefined
-  readonly state: string
-  readonly event: string
-}> {}
+export {
+  /**
+   * Error returned when a state machine does not stabilize within the maximum
+   * number of macrostep iterations.
+   *
+   * @category errors
+   * @since 4.0.0
+   */
+  InfiniteTransitionError,
+  /**
+   * Error returned when a state machine fails while running startup lifecycle
+   * logic after the initial state has been computed.
+   *
+   * @category errors
+   * @since 4.0.0
+   */
+  StartupError,
+  /**
+   * Error returned when an event has no handler for the current state.
+   *
+   * @category errors
+   * @since 4.0.0
+   */
+  UnhandledEventError
+} from "../../internal/stateMachineErrors.ts"
 
-/**
- * Error returned when a state machine does not stabilize within the maximum
- * number of macrostep iterations.
- *
- * @category errors
- * @since 4.0.0
- */
-export class InfiniteTransitionError extends Data.TaggedError("InfiniteTransitionError")<{
-  readonly machineId: string | undefined
-  readonly state: string
-  readonly maxIterations: number
-}> {}
-
-/**
- * Error returned when a state machine fails while running startup lifecycle
- * logic after the initial state has been computed.
- *
- * @category errors
- * @since 4.0.0
- */
-export class StartupError extends Data.TaggedError("StartupError")<{
-  readonly cause: Cause.Cause<unknown>
-}> {}
-
-type DeferredAction<E = any, R = any> = Effect.Effect<void, E, R>
-
-interface DeferredQueue<A> {
-  readonly add: (value: A) => Effect.Effect<void>
-  readonly read: Effect.Effect<ReadonlyArray<A>>
-}
-
-class DeferredActions extends Context.Service<DeferredActions, {
-  readonly add: <E, R>(effect: DeferredAction<E, R>) => Effect.Effect<void>
-  readonly read: Effect.Effect<ReadonlyArray<DeferredAction>>
-}>()("effect/StateMachine/DeferredActions") {}
-
-class DeferredRaisedEvents extends Context.Service<DeferredRaisedEvents, {
-  readonly add: <Event>(event: Event) => Effect.Effect<void>
-  readonly read: Effect.Effect<ReadonlyArray<any>>
-}>()("effect/StateMachine/DeferredRaisedEvents") {}
-
-/**
- * Actor runtime scope available to state machine actions when a machine runs
- * through `toActorLogic`.
- *
- * @category services
- * @since 4.0.0
- */
-export class ActorRuntime extends Context.Service<ActorRuntime, ActorModule.ActorScope<any>>()(
-  "effect/StateMachine/ActorRuntime"
-) {}
+export {
+  /**
+   * Actor runtime scope available to state machine actions when a machine runs
+   * through `toActorLogic`.
+   *
+   * @category services
+   * @since 4.0.0
+   */
+  ActorRuntime
+} from "./internal/stateMachineRuntime.ts"
 
 const RuntimeRequirementTypeId = "~effect/StateMachine/RuntimeRequirement"
 
@@ -213,10 +180,6 @@ export declare namespace Runtime {
     }
   }
 }
-
-class RuntimeContext extends Context.Service<RuntimeContext, Runtime<any, any>>()(
-  "effect/StateMachine/Runtime"
-) {}
 
 type ExcludeCompatibleRuntime<Requirements, Events, Emits> = Requirements extends Runtime.Requirement<
   infer RequiredEvents,
@@ -729,7 +692,7 @@ export declare namespace Machine {
     States extends StateSchemas,
     StateId extends StateIdentifier<States>
   > {
-    readonly [TargetTypeId]: typeof TargetTypeId
+    readonly [Model.TargetTypeId]: typeof Model.TargetTypeId
     readonly path: StateId
     readonly value: StateByIdentifier<States, StateId>
     readonly values?: Partial<
@@ -1326,620 +1289,6 @@ const Proto = {
   }
 }
 
-const getSchemaTag = (schema: Machine.TaggedSchema): PropertyKey | undefined => {
-  const tag = (schema as any).fields?._tag?.schema?.literal ?? (schema as any).fields?._tag?.ast?.literal
-  return typeof tag === "string" || typeof tag === "number" || typeof tag === "symbol" ? tag : undefined
-}
-
-const getStateNodeDefinition = (
-  path: string,
-  definition: Machine.TaggedSchema | Machine.StateNodeConfig
-): {
-  readonly schema: Machine.TaggedSchema
-  readonly type: "atomic" | "compound" | "parallel" | "final"
-  readonly initial: string | undefined
-  readonly states: Machine.StateTree | undefined
-} => {
-  if (Schema.isSchema(definition)) {
-    return { schema: definition as Machine.TaggedSchema, type: "atomic", initial: undefined, states: undefined }
-  }
-  if (!hasProperty(definition, "schema") || !Schema.isSchema(definition.schema)) {
-    throw new Error(`StateMachine.make expected state "${path}" to be a tagged schema or state node config`)
-  }
-  if ((definition as any).type === "parallel" && !hasProperty(definition, "states")) {
-    throw new Error(`StateMachine.make expected parallel state "${path}" to declare child regions`)
-  }
-  if (hasProperty(definition, "states")) {
-    if ((definition as any).type === "final") {
-      throw new Error(`StateMachine.make expected compound state "${path}" to be active`)
-    }
-    if ((definition as any).type === "parallel") {
-      return {
-        schema: definition.schema as Machine.TaggedSchema,
-        type: "parallel",
-        initial: undefined,
-        states: (definition as any).states as Machine.StateTree
-      }
-    }
-    if (typeof (definition as any).initial !== "string") {
-      throw new Error(`StateMachine.make expected compound state "${path}" to declare an initial child`)
-    }
-    return {
-      schema: definition.schema as Machine.TaggedSchema,
-      type: "compound",
-      initial: (definition as any).initial,
-      states: (definition as any).states as Machine.StateTree
-    }
-  }
-  return {
-    schema: definition.schema as Machine.TaggedSchema,
-    type: definition.type === "final" ? "final" : "atomic",
-    initial: undefined,
-    states: undefined
-  }
-}
-
-const compileStateNodes = (states: Machine.StateSchemas): Machine.StateNodes => {
-  const byPath = new Map<string, Machine.StateNode>()
-  let order = 0
-
-  const compile = (tree: Machine.StateTree, parent: string | undefined): ReadonlyArray<string> => {
-    const paths: Array<string> = []
-    for (const key of Object.keys(tree)) {
-      const path = parent === undefined ? key : `${parent}.${key}`
-      const definition = getStateNodeDefinition(path, tree[key])
-      const node = {
-        path,
-        key,
-        tag: getSchemaTag(definition.schema) ?? key,
-        schema: definition.schema,
-        type: definition.type,
-        parent,
-        children: [] as ReadonlyArray<string>,
-        initial: definition.initial === undefined ? undefined : `${path}.${definition.initial}`,
-        order
-      }
-      byPath.set(path, node)
-      paths.push(path)
-      order += 1
-      if (definition.states !== undefined) {
-        const children = compile(definition.states, path)
-        if (node.type === "compound" && (node.initial === undefined || !children.includes(node.initial))) {
-          throw new Error(`StateMachine.make expected compound state "${path}" initial child to exist`)
-        }
-        ;(node as { children: ReadonlyArray<string> }).children = children
-      }
-    }
-    return paths
-  }
-
-  return {
-    byPath,
-    roots: compile(states, undefined)
-  }
-}
-
-const makeTarget: Machine.TargetFunction<any> = (
-  path: string,
-  value: { readonly _tag: PropertyKey },
-  options?: {
-    readonly values?: Readonly<Record<string, unknown>>
-  }
-) =>
-  ({
-    [TargetTypeId]: TargetTypeId,
-    path,
-    value,
-    values: options?.values
-  }) as any
-
-const isTarget = (u: unknown): u is Machine.Target<any, any> => hasProperty(u, TargetTypeId)
-
-const isSnapshot = (u: unknown): u is Machine.AtomicSnapshot<string, unknown> =>
-  hasProperty(u, "path") && hasProperty(u, "value")
-
-const findStateNode = (
-  machine: Machine.Any,
-  value: { readonly _tag: PropertyKey }
-): Machine.StateNode | undefined => {
-  for (const node of machine.stateNodes.byPath.values()) {
-    if (Schema.is(node.schema)(value)) {
-      return node
-    }
-  }
-  return typeof value._tag === "string" ? machine.stateNodes.byPath.get(value._tag) : undefined
-}
-
-interface ActiveConfiguration {
-  readonly active: ReadonlySet<string>
-  readonly values: ReadonlyMap<string, unknown>
-  readonly outputs: ReadonlyMap<string, unknown>
-}
-
-type SnapshotWithCompletedOutputs = Machine.AtomicSnapshot<string, unknown> & {
-  readonly [CompletedOutputsTypeId]?: ReadonlyMap<string, unknown>
-}
-
-interface FinalCompletion {
-  readonly path: string
-  readonly output: unknown
-}
-
-const getNode = (machine: Machine.Any, path: string): Machine.StateNode => {
-  const node = machine.stateNodes.byPath.get(path)
-  if (node === undefined) {
-    throw new Error(`StateMachine expected state path "${path}" to exist`)
-  }
-  return node
-}
-
-const hasOwn = (u: object, key: string): boolean => Object.prototype.hasOwnProperty.call(u, key)
-
-const isDescendantOf = (path: string, ancestor: string): boolean => path.startsWith(`${ancestor}.`)
-
-const isPathInSubtree = (path: string, ancestor: string): boolean => path === ancestor || isDescendantOf(path, ancestor)
-
-const getPathToRoot = (machine: Machine.Any, path: string): ReadonlyArray<string> => {
-  const paths: Array<string> = []
-  let current: string | undefined = path
-  while (current !== undefined) {
-    paths.unshift(current)
-    current = getNode(machine, current).parent
-  }
-  return paths
-}
-
-const pathDepth = (machine: Machine.Any, path: string): number => getPathToRoot(machine, path).length
-
-const compareDocumentOrder = (machine: Machine.Any, left: string, right: string): number =>
-  getNode(machine, left).order - getNode(machine, right).order
-
-const hasActiveChild = (machine: Machine.Any, configuration: ActiveConfiguration, path: string): boolean =>
-  getNode(machine, path).children.some((child) => configuration.active.has(child))
-
-const getActiveLeafPaths = (machine: Machine.Any, configuration: ActiveConfiguration): ReadonlyArray<string> => {
-  const leaves = Array.from(configuration.active)
-    .filter((path) => !hasActiveChild(machine, configuration, path))
-    .sort((left, right) => compareDocumentOrder(machine, left, right))
-  if (leaves.length === 0) {
-    throw new Error("StateMachine expected an active leaf state")
-  }
-  return leaves
-}
-
-const getLeafPath = (machine: Machine.Any, configuration: ActiveConfiguration): string =>
-  getActiveLeafPaths(
-    machine,
-    configuration
-  )[0]
-
-const getActiveLeafPathFrom = (
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  path: string
-): string => {
-  const leaves = getActiveLeafPaths(machine, configuration)
-    .filter((leaf) => isPathInSubtree(leaf, path))
-  if (leaves.length === 0) {
-    throw new Error(`StateMachine expected state "${path}" to have an active leaf state`)
-  }
-  return leaves[0]
-}
-
-const getRootPath = (machine: Machine.Any, configuration: ActiveConfiguration): string => {
-  for (const path of configuration.active) {
-    if (getNode(machine, path).parent === undefined) {
-      return path
-    }
-  }
-  throw new Error("StateMachine expected an active root state")
-}
-
-const getActiveValue = (configuration: ActiveConfiguration, path: string): unknown => {
-  if (!configuration.values.has(path)) {
-    throw new Error(`StateMachine expected active state "${path}" to have a value`)
-  }
-  return configuration.values.get(path)
-}
-
-const getInitialEntryPaths = (
-  machine: Machine.Any,
-  configuration: ActiveConfiguration
-): ReadonlyArray<string> => {
-  const visit = (path: string): ReadonlyArray<string> => {
-    if (!configuration.active.has(path)) {
-      return []
-    }
-    const node = getNode(machine, path)
-    return [
-      path,
-      ...node.children.flatMap(visit)
-    ]
-  }
-  return machine.stateNodes.roots.flatMap(visit)
-}
-
-const snapshotFromPath = <const States extends Machine.StateSchemas>(
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  path: string
-): Machine.SnapshotByIdentifier<States, Machine.StateIdentifier<States>> => {
-  const node = getNode(machine, path)
-  const snapshot: Record<string, unknown> = {
-    path,
-    value: getActiveValue(configuration, path)
-  }
-  if (node.type === "compound") {
-    const child = node.children.find((child) => configuration.active.has(child))
-    if (child === undefined) {
-      throw new Error(`StateMachine expected compound state "${path}" to have an active child`)
-    }
-    snapshot.state = snapshotFromPath(machine, configuration, child)
-  }
-  if (node.type === "parallel") {
-    const states: Record<string, unknown> = {}
-    for (const child of node.children) {
-      if (!configuration.active.has(child)) {
-        throw new Error(`StateMachine expected parallel state "${path}" to have active child region "${child}"`)
-      }
-      const childNode = getNode(machine, child)
-      states[childNode.key] = snapshotFromPath(machine, configuration, child)
-    }
-    snapshot.states = states
-  }
-  return snapshot as unknown as Machine.SnapshotByIdentifier<States, Machine.StateIdentifier<States>>
-}
-
-const snapshotFromConfiguration = <const States extends Machine.StateSchemas>(
-  machine: Machine.Any,
-  configuration: ActiveConfiguration
-): Machine.Snapshot<States> => {
-  const snapshot = snapshotFromPath<States>(
-    machine,
-    configuration,
-    getRootPath(machine, configuration)
-  ) as Machine.Snapshot<States>
-  Object.defineProperty(snapshot, CompletedOutputsTypeId, {
-    value: new Map(configuration.outputs),
-    enumerable: false
-  })
-  return snapshot
-}
-
-const configurationFromSnapshot = (
-  machine: Machine.Any,
-  snapshot: Machine.AtomicSnapshot<string, unknown>
-): ActiveConfiguration => {
-  const active = new Set<string>()
-  const values = new Map<string, unknown>()
-  const snapshotOutputs = (snapshot as SnapshotWithCompletedOutputs)[CompletedOutputsTypeId]
-
-  const visit = (current: Machine.AtomicSnapshot<string, unknown>): void => {
-    const node = getNode(machine, String(current.path))
-    if (!Schema.is(node.schema)(current.value)) {
-      throw new Error(`StateMachine expected snapshot for "${node.path}" to match its schema`)
-    }
-    active.add(node.path)
-    values.set(node.path, current.value)
-    if (node.type === "compound") {
-      if (!hasProperty(current, "state") || !isSnapshot(current.state)) {
-        throw new Error(`StateMachine expected compound snapshot "${node.path}" to include an active child state`)
-      }
-      const child = getNode(machine, String(current.state.path))
-      if (child.parent !== node.path) {
-        throw new Error(`StateMachine expected snapshot "${child.path}" to be a child of "${node.path}"`)
-      }
-      visit(current.state)
-    }
-    if (node.type === "parallel") {
-      if (!hasProperty(current, "states") || typeof current.states !== "object" || current.states === null) {
-        throw new Error(`StateMachine expected parallel snapshot "${node.path}" to include active child regions`)
-      }
-      const states = current.states as Readonly<Record<string, unknown>>
-      for (const childPath of node.children) {
-        const child = getNode(machine, childPath)
-        const childSnapshot = states[child.key]
-        if (!hasOwn(states, child.key) || !isSnapshot(childSnapshot)) {
-          throw new Error(`StateMachine expected parallel snapshot "${node.path}" to include region "${child.key}"`)
-        }
-        const snapshotChild = getNode(machine, String(childSnapshot.path))
-        if (snapshotChild.path !== child.path) {
-          throw new Error(`StateMachine expected snapshot "${snapshotChild.path}" to be region "${child.path}"`)
-        }
-        visit(childSnapshot)
-      }
-    }
-  }
-
-  visit(snapshot)
-  const outputs = new Map<string, unknown>()
-  if (snapshotOutputs !== undefined) {
-    for (const [path, output] of snapshotOutputs) {
-      if (active.has(path)) {
-        outputs.set(path, output)
-      }
-    }
-  }
-  return { active, values, outputs }
-}
-
-const configurationFromValue = (
-  machine: Machine.Any,
-  value: { readonly _tag: PropertyKey }
-): ActiveConfiguration => {
-  const node = findStateNode(machine, value)
-  if (node === undefined) {
-    throw new Error(`StateMachine expected state "${String(value._tag)}" to match a state node`)
-  }
-  if (node.parent !== undefined || node.type === "compound" || node.type === "parallel") {
-    throw new Error(`StateMachine expected state "${node.path}" to be provided as a snapshot with ancestor values`)
-  }
-  return {
-    active: new Set([node.path]),
-    values: new Map([[node.path, value]]),
-    outputs: new Map()
-  }
-}
-
-const normalizeConfiguration = <const States extends Machine.StateSchemas>(
-  machine: Machine.Any,
-  state: Machine.StateLike<States>
-): ActiveConfiguration => {
-  if (isSnapshot(state)) {
-    return configurationFromSnapshot(machine, state)
-  }
-  return configurationFromValue(machine, state)
-}
-
-const validateInitialConfiguration = (machine: Machine.Any, configuration: ActiveConfiguration): void => {
-  for (const path of configuration.active) {
-    const node = getNode(machine, path)
-    if (node.type === "compound") {
-      const child = node.children.find((child) => configuration.active.has(child))
-      if (child !== node.initial) {
-        throw new Error(`StateMachine initial state "${node.path}" must enter initial child "${node.initial}"`)
-      }
-    }
-    if (node.type === "parallel") {
-      for (const child of node.children) {
-        if (!configuration.active.has(child)) {
-          throw new Error(`StateMachine initial state "${node.path}" must enter child region "${child}"`)
-        }
-      }
-    }
-  }
-}
-
-const configurationFromTargetPath = (
-  machine: Machine.Any,
-  current: ActiveConfiguration,
-  path: string,
-  value: { readonly _tag: PropertyKey },
-  providedValues: Readonly<Record<string, unknown>> | undefined
-): ActiveConfiguration => {
-  const node = getNode(machine, path)
-  const active = new Set<string>()
-  const values = new Map<string, unknown>()
-  const outputs = new Map<string, unknown>()
-  const paths = getPathToRoot(machine, node.path)
-  const pathSet = new Set(paths)
-
-  for (const currentPath of paths) {
-    active.add(currentPath)
-    if (currentPath === node.path) {
-      values.set(currentPath, value)
-    } else if (providedValues !== undefined && hasOwn(providedValues, currentPath)) {
-      values.set(currentPath, providedValues[currentPath])
-    } else if (current.values.has(currentPath)) {
-      values.set(currentPath, current.values.get(currentPath))
-    } else {
-      throw new Error(`StateMachine target "${node.path}" requires a value for ancestor state "${currentPath}"`)
-    }
-  }
-
-  for (const ancestor of paths) {
-    const ancestorNode = getNode(machine, ancestor)
-    if (ancestorNode.type === "parallel") {
-      for (const child of ancestorNode.children) {
-        if (pathSet.has(child) || !current.active.has(child)) {
-          continue
-        }
-        for (const activePath of current.active) {
-          if (isPathInSubtree(activePath, child)) {
-            active.add(activePath)
-            if (current.values.has(activePath)) {
-              values.set(activePath, current.values.get(activePath))
-            }
-            if (current.outputs.has(activePath)) {
-              outputs.set(activePath, current.outputs.get(activePath))
-            }
-          }
-        }
-      }
-    }
-  }
-
-  if (node.type === "compound" || node.type === "parallel") {
-    throw new Error(`StateMachine target "${node.path}" must include an active child state`)
-  }
-
-  return { active, values, outputs }
-}
-
-const normalizeTargetConfiguration = <const States extends Machine.StateSchemas>(
-  machine: Machine.Any,
-  current: ActiveConfiguration,
-  target: Machine.StateOf<States> | Machine.Snapshot<States> | Machine.Target<States, Machine.StateIdentifier<States>>
-): ActiveConfiguration => {
-  if (isTarget(target)) {
-    return configurationFromTargetPath(
-      machine,
-      current,
-      target.path,
-      target.value as { readonly _tag: PropertyKey },
-      target.values as Readonly<Record<string, unknown>> | undefined
-    )
-  }
-  if (isSnapshot(target)) {
-    return configurationFromSnapshot(machine, target)
-  }
-  const node = findStateNode(machine, target as { readonly _tag: PropertyKey })
-  if (node === undefined) {
-    throw new Error(`StateMachine expected target state "${String((target as any)._tag)}" to match a state node`)
-  }
-  return configurationFromTargetPath(machine, current, node.path, target as { readonly _tag: PropertyKey }, undefined)
-}
-
-const getStateConfigByPath = (
-  machine: Machine.Any,
-  path: string
-): Machine.AnyStateConfig | undefined => machine.handlers[path]
-
-const getActiveChildPath = (
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  path: string
-): string | undefined => getNode(machine, path).children.find((child) => configuration.active.has(child))
-
-const isDirectFinalPath = (
-  machine: Machine.Any,
-  path: string
-): boolean => getNode(machine, path).type === "final" || getStateConfigByPath(machine, path)?.type === "final"
-
-const isActiveFinalNode = (
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  path: string
-): boolean => {
-  if (!configuration.active.has(path)) {
-    return false
-  }
-  const node = getNode(machine, path)
-  if (node.type === "compound") {
-    const child = getActiveChildPath(machine, configuration, path)
-    return child !== undefined && isActiveFinalNode(machine, configuration, child)
-  }
-  if (node.type === "parallel") {
-    for (const child of node.children) {
-      if (!isActiveFinalNode(machine, configuration, child)) {
-        return false
-      }
-    }
-    return true
-  }
-  return isDirectFinalPath(machine, path)
-}
-
-const isActiveFinalConfiguration = (
-  machine: Machine.Any,
-  configuration: ActiveConfiguration
-): boolean => isActiveFinalNode(machine, configuration, getRootPath(machine, configuration))
-
-const setCompletedOutput = (
-  outputs: Map<string, unknown>,
-  path: string,
-  output: unknown
-): unknown => {
-  if (!outputs.has(path)) {
-    outputs.set(path, output)
-  }
-  return outputs.get(path)
-}
-
-const resolveFinalOutput = <const Events extends ReadonlyArray<Machine.TaggedSchema>>(
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  path: string,
-  event: Machine.EventOf<Events>,
-  outputs?: Readonly<Record<string, unknown>>
-): unknown =>
-  getStateConfigByPath(machine, path)?.output?.({
-    state: getActiveValue(configuration, path),
-    event,
-    outputs
-  } as any)
-
-const completeActiveFinalNode = <const Events extends ReadonlyArray<Machine.TaggedSchema>>(
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  path: string,
-  event: Machine.EventOf<Events>,
-  outputs: Map<string, unknown>
-): FinalCompletion | undefined => {
-  if (!configuration.active.has(path)) {
-    return undefined
-  }
-  const node = getNode(machine, path)
-  if (node.type === "compound") {
-    const child = getActiveChildPath(machine, configuration, path)
-    if (child === undefined) {
-      return undefined
-    }
-    const childCompletion = completeActiveFinalNode(machine, configuration, child, event, outputs)
-    if (childCompletion === undefined) {
-      return undefined
-    }
-    return {
-      path,
-      output: setCompletedOutput(outputs, path, childCompletion.output)
-    }
-  }
-  if (node.type === "parallel") {
-    const regionOutputs: Record<string, unknown> = {}
-    let completed = true
-    for (const child of node.children) {
-      const childCompletion = completeActiveFinalNode(machine, configuration, child, event, outputs)
-      if (childCompletion === undefined) {
-        completed = false
-      } else {
-        regionOutputs[getNode(machine, child).key] = childCompletion.output
-      }
-    }
-    if (!completed) {
-      return undefined
-    }
-    return {
-      path,
-      output: setCompletedOutput(
-        outputs,
-        path,
-        resolveFinalOutput(machine, configuration, path, event, regionOutputs)
-      )
-    }
-  }
-  if (!isDirectFinalPath(machine, path)) {
-    return undefined
-  }
-  return {
-    path,
-    output: setCompletedOutput(outputs, path, resolveFinalOutput(machine, configuration, path, event))
-  }
-}
-
-const completeConfiguration = <const Events extends ReadonlyArray<Machine.TaggedSchema>>(
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  event: Machine.EventOf<Events>
-): {
-  readonly configuration: ActiveConfiguration
-  readonly completion: FinalCompletion | undefined
-} => {
-  const outputs = new Map(configuration.outputs)
-  const completed = {
-    active: configuration.active,
-    values: configuration.values,
-    outputs
-  }
-  const completion = completeActiveFinalNode(
-    machine,
-    completed,
-    getRootPath(machine, completed),
-    event,
-    outputs
-  )
-  return { configuration: completed, completion }
-}
-
 const handleUnsafe = (
   self: Machine.Any,
   stateTag: PropertyKey,
@@ -1960,633 +1309,6 @@ const handleUnsafe = (
   return machine
 }
 
-const makeDeferredQueue = <A>(): Effect.Effect<DeferredQueue<A>> =>
-  Effect.sync(() => {
-    const values: Array<A> = []
-    return {
-      read: Effect.sync(() => values),
-      add: (value) =>
-        Effect.sync(() => {
-          values.push(value)
-        })
-    }
-  })
-
-const makeDeferredActions = Effect.map(
-  makeDeferredQueue<DeferredAction>(),
-  (queue) =>
-    DeferredActions.of({
-      read: queue.read,
-      add: (effect) => queue.add(effect)
-    })
-)
-
-const makeDeferredRaisedEvents = Effect.map(
-  makeDeferredQueue<any>(),
-  (queue) =>
-    DeferredRaisedEvents.of({
-      read: queue.read,
-      add: (event) => queue.add(event)
-    })
-)
-
-const provideDeferredServices = <A, E, R>(
-  effect: Effect.Effect<A, E, R>,
-  deferredActions: DeferredActions["Service"],
-  deferredRaisedEvents: DeferredRaisedEvents["Service"]
-): Effect.Effect<A, E, R> =>
-  effect.pipe(
-    Effect.provideService(DeferredActions, deferredActions),
-    Effect.provideService(DeferredRaisedEvents, deferredRaisedEvents),
-    Effect.provideService(RuntimeContext, makePlanningRuntime(deferredRaisedEvents))
-  )
-
-const provideActorRuntime = <A, E, R, Event>(
-  effect: Effect.Effect<A, E, R>,
-  scope: ActorModule.ActorScope<Event>
-): Effect.Effect<A, E, Exclude<R, ActorRuntime>> =>
-  Effect.provideService(effect, ActorRuntime, scope as ActorModule.ActorScope<any>)
-
-const provideRuntimeContext = <A, E, R, Events, Emits>(
-  effect: Effect.Effect<A, E, R>,
-  runtime: Runtime<Events, Emits>
-): Effect.Effect<A, E, R> =>
-  Effect.provideService(
-    effect as Effect.Effect<A, E, R | RuntimeContext>,
-    RuntimeContext,
-    runtime as Runtime<any, any>
-  ) as Effect.Effect<A, E, R>
-
-const sendParentOptional = <Event>(event: Event): Effect.Effect<void> =>
-  Effect.contextWith((context: Context.Context<never>) => {
-    const runtime = Context.getOption(context as Context.Context<ActorRuntime>, ActorRuntime)
-    return Option.isSome(runtime) && runtime.value.parent !== undefined
-      ? runtime.value.parent.send(event)
-      : Effect.void
-  })
-
-const makePlanningRuntime = <Events, Emits>(
-  deferredRaisedEvents: DeferredRaisedEvents["Service"]
-): Runtime<Events, Emits> =>
-  RuntimeContext.of({
-    raise: (event) => deferredRaisedEvents.add(event),
-    sendParent: sendParentOptional
-  })
-
-const makeActorRuntime = <Events, Emits>(
-  scope: ActorModule.ActorScope<Events>
-): Runtime<Events, Emits> =>
-  RuntimeContext.of({
-    raise: (event) => scope.self.send(event),
-    sendParent: (event) => scope.parent === undefined ? Effect.void : scope.parent.send(event)
-  })
-
-const runActions = <E, R, Events, Emits>(
-  actions: Iterable<Effect.Effect<void, E, R>>,
-  runtime: Runtime<Events, Emits>
-): Effect.Effect<void, E, R> =>
-  Effect.all(
-    Array.from(actions, (action) => provideRuntimeContext(action, runtime)),
-    { discard: true }
-  )
-
-const runtimeFor = <Events, Emits>(): Effect.Effect<
-  Runtime<Events, Emits>,
-  never,
-  Runtime.Requirement<Events, Emits>
-> => runtime<{ readonly events: Events; readonly emits: Emits }>()
-
-const runStateAction = <Context, E, R>(
-  handler: ((context: Context) => Machine.StateActionResult<E, R>) | undefined,
-  context: Context,
-  deferredActions: DeferredActions["Service"],
-  deferredRaisedEvents: DeferredRaisedEvents["Service"]
-): Effect.Effect<void, E, R> => {
-  if (handler === undefined) {
-    return Effect.void
-  }
-
-  const result = handler(context)
-  return Effect.isEffect(result)
-    ? provideDeferredServices(result, deferredActions, deferredRaisedEvents)
-    : Effect.void
-}
-
-type MicrostepPlan<State, Event, E, R> = {
-  readonly next: State
-  readonly actions: ReadonlyArray<Effect.Effect<void, E, R>>
-  readonly raisedEvents: ReadonlyArray<Event>
-  readonly exitPaths: ReadonlyArray<string>
-  readonly entryPaths: ReadonlyArray<string>
-  readonly changed: boolean
-}
-
-type MacrostepPlan<State, Event, E, R, Output> = {
-  readonly next: State
-  readonly actions: ReadonlyArray<Effect.Effect<void, E, R>>
-  readonly microsteps: ReadonlyArray<MicrostepPlan<State, Event, E, R>>
-  readonly output: Output | undefined
-}
-
-type TransitionHandler<States extends Machine.StateSchemas, E, R, Context> = (
-  context: Context
-) => Machine.HandlerResult<States, E, R>
-
-type EventTransition<States extends Machine.StateSchemas, E, R, Context> =
-  | TransitionHandler<States, E, R, Context>
-  | {
-    readonly reenter?: boolean
-    readonly transition: TransitionHandler<States, E, R, Context>
-  }
-
-type MicrostepTransition<States extends Machine.StateSchemas, E, R, Context> = {
-  readonly reenter: boolean
-  readonly transition: TransitionHandler<States, E, R, Context>
-}
-
-type AnyInvokeConfig = Machine.InvokeConfig<any, any, any, any, any, any, any, any, any, any, any>
-
-interface InvokeSession {
-  readonly token: symbol
-  readonly scope: Scope.Closeable
-  readonly childId: string
-}
-
-const normalizeEventTransition = <States extends Machine.StateSchemas, E, R, Context>(
-  transition: EventTransition<States, E, R, Context> | undefined
-): MicrostepTransition<States, E, R, Context> | undefined => {
-  if (transition === undefined) {
-    return undefined
-  }
-  return typeof transition === "function"
-    ? { reenter: false, transition }
-    : { reenter: transition.reenter === true, transition: transition.transition }
-}
-
-const getInvokes = (config: Machine.AnyStateConfig | undefined): ReadonlyArray<AnyInvokeConfig> => {
-  const invokes = config?.invoke
-  if (invokes === undefined) {
-    return []
-  }
-  return Array.isArray(invokes) ? invokes as ReadonlyArray<AnyInvokeConfig> : [invokes as AnyInvokeConfig]
-}
-
-const collectStateAction = Effect.fnUntraced(function*<Context, Event, E, R>(
-  handler: ((context: Context) => Machine.StateActionResult<E, R>) | undefined,
-  context: Context
-) {
-  const deferredActions = yield* makeDeferredActions
-  const deferredRaisedEvents = yield* makeDeferredRaisedEvents
-  yield* runStateAction(handler, context, deferredActions, deferredRaisedEvents)
-  const actions = yield* deferredActions.read
-  const raisedEvents = yield* deferredRaisedEvents.read
-  return {
-    actions: actions as ReadonlyArray<DeferredAction<E, R>>,
-    raisedEvents: raisedEvents as ReadonlyArray<Event>
-  }
-})
-
-const collectTransition = Effect.fnUntraced(function*<
-  const States extends Machine.StateSchemas,
-  Event,
-  E,
-  R,
-  Context
->(
-  transition: TransitionHandler<States, E, R, Context>,
-  context: Context
-) {
-  const deferredActions = yield* makeDeferredActions
-  const deferredRaisedEvents = yield* makeDeferredRaisedEvents
-  const result = transition(context)
-  const state = Effect.isEffect(result)
-    ? yield* provideDeferredServices(result, deferredActions, deferredRaisedEvents)
-    : result
-  const actions = yield* deferredActions.read
-  const raisedEvents = yield* deferredRaisedEvents.read
-  return {
-    state,
-    actions: actions as ReadonlyArray<DeferredAction<E, R>>,
-    raisedEvents: raisedEvents as ReadonlyArray<Event>
-  }
-})
-
-type SelectedTransition<States extends Machine.StateSchemas, E, R, Context> = {
-  readonly sourcePath: string
-  readonly leafPath: string
-  readonly transition: MicrostepTransition<States, E, R, Context>
-  readonly context: Context
-}
-
-type EvaluatedTransition<States extends Machine.StateSchemas, Event, E, R, Context> = {
-  readonly selection: SelectedTransition<States, E, R, Context>
-  readonly target:
-    | Machine.StateOf<States>
-    | Machine.Snapshot<States>
-    | Machine.Target<States, Machine.StateIdentifier<States>>
-    | undefined
-  readonly actions: ReadonlyArray<Effect.Effect<void, E, R>>
-  readonly raisedEvents: ReadonlyArray<Event>
-  readonly changed: boolean
-  readonly exitPaths: ReadonlyArray<string>
-  readonly entryPaths: ReadonlyArray<string>
-}
-
-const getCandidatePaths = (machine: Machine.Any, configuration: ActiveConfiguration): ReadonlyArray<string> =>
-  Array.from(configuration.active)
-    .sort((left, right) => {
-      const depth = pathDepth(machine, right) - pathDepth(machine, left)
-      return depth === 0 ? compareDocumentOrder(machine, left, right) : depth
-    })
-
-const getLeafCandidatePaths = (machine: Machine.Any, leaf: string): ReadonlyArray<string> =>
-  [...getPathToRoot(machine, leaf)].reverse()
-
-const getLeastCommonAncestor = (
-  machine: Machine.Any,
-  left: string,
-  right: string
-): string | undefined => {
-  const leftPath = getPathToRoot(machine, left)
-  const rightPath = getPathToRoot(machine, right)
-  let ancestor: string | undefined = undefined
-  const length = Math.min(leftPath.length, rightPath.length)
-  for (let index = 0; index < length; index++) {
-    if (leftPath[index] !== rightPath[index]) {
-      break
-    }
-    ancestor = leftPath[index]
-  }
-  return ancestor
-}
-
-const getExitPaths = (
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  boundary: string | undefined
-): ReadonlyArray<string> =>
-  sortExitPaths(
-    machine,
-    Array.from(configuration.active)
-      .filter((path) => boundary === undefined || isDescendantOf(path, boundary))
-  )
-
-const getEntryPaths = (
-  machine: Machine.Any,
-  targetLeaf: string,
-  boundary: string | undefined
-): ReadonlyArray<string> =>
-  getPathToRoot(machine, targetLeaf).filter((path) => boundary === undefined || isDescendantOf(path, boundary))
-
-const sortExitPaths = (machine: Machine.Any, paths: Iterable<string>): ReadonlyArray<string> =>
-  Array.from(new Set(paths))
-    .sort((left, right) => {
-      const depth = getPathToRoot(machine, right).length - getPathToRoot(machine, left).length
-      return depth === 0 ? getNode(machine, right).order - getNode(machine, left).order : depth
-    })
-
-const sortEntryPaths = (machine: Machine.Any, paths: Iterable<string>): ReadonlyArray<string> =>
-  Array.from(new Set(paths))
-    .sort((left, right) => {
-      const depth = getPathToRoot(machine, left).length - getPathToRoot(machine, right).length
-      return depth === 0 ? compareDocumentOrder(machine, left, right) : depth
-    })
-
-const makeStateActionContext = <
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema>,
-  StateId extends Machine.StateIdentifier<States>
->(
-  configuration: ActiveConfiguration,
-  path: string,
-  event: Machine.EventOf<Events>
-): Machine.StateActionContext<States, Events, Emits, StateId> => ({
-  state: getActiveValue(configuration, path) as Machine.StateByIdentifier<States, StateId>,
-  event,
-  runtime: runtimeFor<Machine.EventOf<Events>, Machine.EmitOf<Emits>>()
-})
-
-const makeTransitionContext = <
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema>,
-  StateId extends Machine.StateIdentifier<States>,
-  EventTag extends Machine.TagOf<Events[number]>
->(
-  configuration: ActiveConfiguration,
-  path: string,
-  event: Machine.EventByTag<Events, EventTag>
-): Machine.HandlerContext<States, Events, Emits, StateId, EventTag, any, any> => ({
-  state: getActiveValue(configuration, path) as Machine.StateByIdentifier<States, StateId>,
-  event,
-  runtime: runtimeFor<Machine.EventOf<Events>, Machine.EmitOf<Emits>>(),
-  target: makeTarget
-})
-
-const collectStateActions = Effect.fnUntraced(function*<
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema>,
-  E,
-  R
->(
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  paths: ReadonlyArray<string>,
-  event: Machine.EventOf<Events>,
-  key: "entry" | "exit"
-) {
-  const actions: Array<DeferredAction<E, R>> = []
-  const raisedEvents: Array<Machine.EventOf<Events>> = []
-  for (const path of paths) {
-    const collected = yield* collectStateAction<
-      Machine.StateActionContext<States, Events, Emits, Machine.StateIdentifier<States>>,
-      Machine.EventOf<Events>,
-      E,
-      R
-    >(
-      machine.handlers[path]?.[key],
-      makeStateActionContext<States, Events, Emits, Machine.StateIdentifier<States>>(configuration, path, event)
-    )
-    actions.push(...collected.actions)
-    raisedEvents.push(...collected.raisedEvents)
-  }
-  return { actions, raisedEvents }
-})
-
-const selectAlwaysTransitions = <
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema>,
-  E,
-  R
->(
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  event: Machine.EventOf<Events>
-): ReadonlyArray<
-  SelectedTransition<
-    States,
-    E,
-    R,
-    Machine.AlwaysContext<States, Events, Emits, Machine.StateIdentifier<States>>
-  >
-> => {
-  const selected: Array<
-    SelectedTransition<
-      States,
-      E,
-      R,
-      Machine.AlwaysContext<States, Events, Emits, Machine.StateIdentifier<States>>
-    >
-  > = []
-  const selectedSources = new Set<string>()
-  for (const leaf of getActiveLeafPaths(machine, configuration)) {
-    for (const path of getLeafCandidatePaths(machine, leaf)) {
-      const always = machine.handlers[path]?.always
-      if (always !== undefined) {
-        if (!selectedSources.has(path)) {
-          selectedSources.add(path)
-          selected.push({
-            sourcePath: path,
-            leafPath: leaf,
-            transition: { reenter: false, transition: always },
-            context: {
-              state: getActiveValue(configuration, path) as Machine.StateByIdentifier<
-                States,
-                Machine.StateIdentifier<States>
-              >,
-              event,
-              runtime: runtimeFor<Machine.EventOf<Events>, Machine.EmitOf<Emits>>(),
-              target: makeTarget
-            }
-          })
-        }
-        break
-      }
-    }
-  }
-  return selected
-}
-
-const selectEventTransitions = <
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema>,
-  E,
-  R
->(
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  event: Machine.EventByTag<Events, Machine.TagOf<Events[number]>>
-): ReadonlyArray<
-  SelectedTransition<
-    States,
-    E,
-    R,
-    Machine.HandlerContext<States, Events, Emits, Machine.StateIdentifier<States>, Machine.TagOf<Events[number]>, E, R>
-  >
-> => {
-  const selected: Array<
-    SelectedTransition<
-      States,
-      E,
-      R,
-      Machine.HandlerContext<
-        States,
-        Events,
-        Emits,
-        Machine.StateIdentifier<States>,
-        Machine.TagOf<Events[number]>,
-        E,
-        R
-      >
-    >
-  > = []
-  const selectedSources = new Set<string>()
-  for (const leaf of getActiveLeafPaths(machine, configuration)) {
-    for (const path of getLeafCandidatePaths(machine, leaf)) {
-      const transition = normalizeEventTransition(machine.handlers[path]?.on?.[event._tag])
-      if (transition !== undefined) {
-        if (!selectedSources.has(path)) {
-          selectedSources.add(path)
-          selected.push({
-            sourcePath: path,
-            leafPath: leaf,
-            transition: transition as unknown as MicrostepTransition<
-              States,
-              E,
-              R,
-              Machine.HandlerContext<
-                States,
-                Events,
-                Emits,
-                Machine.StateIdentifier<States>,
-                Machine.TagOf<Events[number]>,
-                E,
-                R
-              >
-            >,
-            context: makeTransitionContext<
-              States,
-              Events,
-              Emits,
-              Machine.StateIdentifier<States>,
-              Machine.TagOf<Events[number]>
-            >(configuration, path, event)
-          })
-        }
-        break
-      }
-    }
-  }
-  return selected
-}
-
-const getTargetNodePath = <const States extends Machine.StateSchemas>(
-  machine: Machine.Any,
-  target: Machine.StateOf<States> | Machine.Snapshot<States> | Machine.Target<States, Machine.StateIdentifier<States>>
-): string => {
-  if (isTarget(target)) {
-    return String(target.path)
-  }
-  if (isSnapshot(target)) {
-    return String(target.path)
-  }
-  const node = findStateNode(machine, target as { readonly _tag: PropertyKey })
-  if (node === undefined) {
-    throw new Error(`StateMachine expected target state "${String((target as any)._tag)}" to match a state node`)
-  }
-  return node.path
-}
-
-const hasPathIntersection = (left: ReadonlyArray<string>, right: ReadonlyArray<string>): boolean => {
-  for (const path of left) {
-    if (right.includes(path)) {
-      return true
-    }
-  }
-  return false
-}
-
-const sortEvaluatedTransitions = <
-  const States extends Machine.StateSchemas,
-  Event,
-  E,
-  R,
-  Context
->(
-  machine: Machine.Any,
-  transitions: Iterable<EvaluatedTransition<States, Event, E, R, Context>>
-): ReadonlyArray<EvaluatedTransition<States, Event, E, R, Context>> =>
-  Array.from(transitions)
-    .sort((left, right) => compareDocumentOrder(machine, left.selection.sourcePath, right.selection.sourcePath))
-
-const removeConflictingTransitions = <
-  const States extends Machine.StateSchemas,
-  Event,
-  E,
-  R,
-  Context
->(
-  machine: Machine.Any,
-  transitions: ReadonlyArray<EvaluatedTransition<States, Event, E, R, Context>>
-): ReadonlyArray<EvaluatedTransition<States, Event, E, R, Context>> => {
-  const filtered: Array<EvaluatedTransition<States, Event, E, R, Context>> = []
-  for (const transition of sortEvaluatedTransitions(machine, transitions)) {
-    let preempted = false
-    const transitionsToRemove = new Set<EvaluatedTransition<States, Event, E, R, Context>>()
-    for (const selected of filtered) {
-      if (hasPathIntersection(transition.exitPaths, selected.exitPaths)) {
-        if (isDescendantOf(transition.selection.sourcePath, selected.selection.sourcePath)) {
-          transitionsToRemove.add(selected)
-        } else {
-          preempted = true
-          break
-        }
-      }
-    }
-    if (!preempted) {
-      for (const removed of transitionsToRemove) {
-        const index = filtered.indexOf(removed)
-        if (index >= 0) {
-          filtered.splice(index, 1)
-        }
-      }
-      filtered.push(transition)
-    }
-  }
-  return filtered
-}
-
-const collectEvaluatedTransition = Effect.fnUntraced(function*<
-  const States extends Machine.StateSchemas,
-  Event,
-  E,
-  R,
-  Context
->(
-  machine: Machine.Any,
-  state: ActiveConfiguration,
-  selection: SelectedTransition<States, E, R, Context>
-) {
-  const stateIdentifier = selection.leafPath
-  const transitionResult = yield* collectTransition<States, Event, E, R, Context>(
-    selection.transition.transition,
-    selection.context
-  )
-  const target = transitionResult.state === undefined
-    ? undefined
-    : transitionResult.state as
-      | Machine.StateOf<States>
-      | Machine.Snapshot<States>
-      | Machine.Target<States, Machine.StateIdentifier<States>>
-  const targetPath = target === undefined ? undefined : getTargetNodePath(machine, target)
-  const stateAfterTransition = target === undefined
-    ? state
-    : normalizeTargetConfiguration<States>(machine, state, target)
-  const targetIdentifier = targetPath === undefined
-    ? stateIdentifier
-    : getActiveLeafPathFrom(machine, stateAfterTransition, targetPath)
-  const changed = targetIdentifier !== stateIdentifier || selection.transition.reenter
-
-  if (!changed) {
-    return {
-      selection,
-      target,
-      actions: transitionResult.actions,
-      raisedEvents: transitionResult.raisedEvents,
-      changed,
-      exitPaths: [],
-      entryPaths: []
-    } as EvaluatedTransition<States, Event, E, R, Context>
-  }
-
-  const boundary = selection.transition.reenter
-    ? getNode(machine, selection.sourcePath).parent
-    : getLeastCommonAncestor(machine, stateIdentifier, targetIdentifier)
-
-  return {
-    selection,
-    target,
-    actions: transitionResult.actions,
-    raisedEvents: transitionResult.raisedEvents,
-    changed,
-    exitPaths: getExitPaths(machine, state, boundary),
-    entryPaths: getEntryPaths(machine, targetIdentifier, boundary)
-  } as EvaluatedTransition<States, Event, E, R, Context>
-})
-
-const MaxMacrostepIterations = 1000
-const InitialEventTypeId: unique symbol = Symbol("effect/StateMachine/InitialEvent")
-const InitialEvent = { _tag: InitialEventTypeId }
-
-const catchStartup = <A>(
-  effect: Effect.Effect<A, unknown, unknown>
-): Effect.Effect<A, StartupError> =>
-  Effect.catchCause(effect as Effect.Effect<A, unknown, never>, (cause) => Effect.fail(new StartupError({ cause })))
-
 /**
  * Returns `true` if a value is a `StateMachine`.
  *
@@ -2596,38 +1318,6 @@ const catchStartup = <A>(
 export const isMachine = (
   u: unknown
 ): u is Machine.Any => hasProperty(u, TypeId)
-
-const isFinalState = (
-  machine: Machine.Any,
-  state: Machine.StateLike<any>
-): boolean => isActiveFinalConfiguration(machine, normalizeConfiguration(machine, state))
-
-const getFinalOutputFromConfiguration = <
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  Output
->(
-  machine: Machine.Any,
-  configuration: ActiveConfiguration,
-  event: Machine.EventOf<Events>
-): Output | undefined => {
-  const completed = completeConfiguration(machine, configuration, event)
-  return completed.completion?.output as Output | undefined
-}
-
-const getFinalOutput = <
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  Output
->(
-  machine: Machine.Any,
-  state: Machine.StateLike<States>,
-  event: Machine.EventOf<Events>
-): Output | undefined =>
-  getFinalOutputFromConfiguration<Events, Output>(
-    machine,
-    normalizeConfiguration(machine, state),
-    event
-  )
 
 /**
  * Returns `true` if a state is final for a state machine.
@@ -2651,7 +1341,7 @@ export const isFinal = <
 ): state is Extract<
   Machine.StateLike<States>,
   Machine.StateByIdentifier<States, FinalStates> | Machine.SnapshotContainingFinal<States, FinalStates>
-> => isFinalState(machine, state)
+> => internalRuntime.isFinal(machine, state)
 
 /**
  * Creates a schema-first state machine definition.
@@ -2695,7 +1385,7 @@ export const make = <
   self.input = config.input
   self.id = config.id
   self.initial = config.initial
-  self.stateNodes = compileStateNodes(config.states)
+  self.stateNodes = Model.compileStateNodes(config.states)
   self.handlers = {}
   return self
 }
@@ -2782,59 +1472,7 @@ export const planInitial: <
   },
   InitialE | StartupError,
   never
-> = Effect.fnUntraced(function*<
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema> = any,
-  const Input extends Schema.Top = typeof Schema.Void,
-  UnhandledStates extends Machine.StateIdentifier<States> = Machine.StateIdentifier<States>,
-  E = never,
-  R = never,
-  InitialE = never,
-  InitialR = never,
-  FinalStates extends Machine.StateIdentifier<States> = never,
-  Output = never
->(
-  machine: Machine<States, Events, Input, UnhandledStates, E, R, InitialE, InitialR, FinalStates, Output, Emits>,
-  ...args: [...Machine.InputArgs<Input>]
-) {
-  const deferredActions = yield* makeDeferredActions
-  const result = machine.initial(...args)
-  const state = Effect.isEffect(result)
-    ? yield* (result.pipe(Effect.provideService(DeferredActions, deferredActions)) as Effect.Effect<
-      Machine.StateOf<States> | Machine.Snapshot<States>
-    >)
-    : result
-  const configuration = normalizeConfiguration<States>(machine, state)
-  validateInitialConfiguration(machine, configuration)
-  const actions = yield* deferredActions.read
-  const settled = yield* catchStartup(Effect.gen(function*() {
-    const entry = yield* collectStateActions<States, Events, Emits, never, never>(
-      machine,
-      configuration,
-      getInitialEntryPaths(machine, configuration),
-      InitialEvent as Machine.EventOf<Events>,
-      "entry"
-    )
-    return yield* (settle(
-      machine,
-      configuration,
-      InitialEvent as Machine.EventOf<Events>,
-      [...entry.actions] as Array<Effect.Effect<void, never, never>>,
-      [...entry.raisedEvents] as Array<Machine.EventOf<Events>>,
-      []
-    ) as Effect.Effect<MacrostepPlan<ActiveConfiguration, Machine.EventOf<Events>, never, never, Output>>)
-  }))
-
-  return {
-    state: snapshotFromConfiguration<States>(machine, settled.next),
-    actions: [
-      ...actions,
-      ...settled.actions.map((action) => catchStartup(action))
-    ] as ReadonlyArray<Effect.Effect<void, InitialE | StartupError, InitialR>>,
-    output: settled.output
-  }
-})
+> = internalRuntime.planInitial
 
 /**
  * Returns the event tags handled by the current state.
@@ -2852,333 +1490,7 @@ export const enabled = <
 >(
   machine: Machine<States, Events, Input, UnhandledStates, E, R>,
   state: Machine.StateLike<States>
-): ReadonlyArray<Machine.TagOf<Events[number]>> => {
-  if (isFinalState(machine, state)) {
-    return []
-  }
-  const configuration = normalizeConfiguration(machine, state)
-  const tags: Array<Machine.TagOf<Events[number]>> = []
-  const seen = new Set<PropertyKey>()
-  for (const path of getCandidatePaths(machine, configuration)) {
-    for (const tag of Reflect.ownKeys(machine.handlers[path]?.on ?? {})) {
-      if (!seen.has(tag)) {
-        seen.add(tag)
-        tags.push(tag as Machine.TagOf<Events[number]>)
-      }
-    }
-  }
-  return tags
-}
-
-const microstep: <
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema> = any,
-  const Input extends Schema.Top = typeof Schema.Void,
-  UnhandledStates extends Machine.StateIdentifier<States> = Machine.StateIdentifier<States>,
-  E = never,
-  R = never,
-  InitialE = never,
-  InitialR = never,
-  FinalStates extends Machine.StateIdentifier<States> = never,
-  Output = never,
-  Context = never
->(
-  machine: Machine<States, Events, Input, UnhandledStates, E, R, InitialE, InitialR, FinalStates, Output, Emits>,
-  state: ActiveConfiguration,
-  event: Machine.EventOf<Events>,
-  selections: ReadonlyArray<SelectedTransition<States, E, R, Context>>
-) => Effect.Effect<
-  MicrostepPlan<ActiveConfiguration, Machine.EventOf<Events>, E, R>,
-  E | UnhandledEventError,
-  R
-> = Effect.fnUntraced(function*<
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema> = any,
-  const Input extends Schema.Top = typeof Schema.Void,
-  UnhandledStates extends Machine.StateIdentifier<States> = Machine.StateIdentifier<States>,
-  E = never,
-  R = never,
-  InitialE = never,
-  InitialR = never,
-  FinalStates extends Machine.StateIdentifier<States> = never,
-  Output = never,
-  Context = never
->(
-  machine: Machine<States, Events, Input, UnhandledStates, E, R, InitialE, InitialR, FinalStates, Output, Emits>,
-  state: ActiveConfiguration,
-  event: Machine.EventOf<Events>,
-  selections: ReadonlyArray<SelectedTransition<States, E, R, Context>>
-) {
-  if (selections.length === 0) {
-    return yield* new UnhandledEventError({
-      machineId: machine.id,
-      state: String(getLeafPath(machine, state)),
-      event: String(event._tag)
-    })
-  }
-
-  const evaluatedTransitions: Array<EvaluatedTransition<States, Machine.EventOf<Events>, E, R, Context>> = []
-  for (const selection of selections) {
-    evaluatedTransitions.push(
-      yield* collectEvaluatedTransition<States, Machine.EventOf<Events>, E, R, Context>(
-        machine,
-        state,
-        selection
-      )
-    )
-  }
-
-  const transitions = removeConflictingTransitions(machine, evaluatedTransitions)
-  let stateAfterTransition = state
-  for (const transition of sortEvaluatedTransitions(machine, transitions)) {
-    if (transition.target !== undefined) {
-      stateAfterTransition = normalizeTargetConfiguration<States>(machine, stateAfterTransition, transition.target)
-    }
-  }
-
-  const changed = transitions.some((transition) => transition.changed)
-  const transitionActions = sortEvaluatedTransitions(machine, transitions)
-    .flatMap((transition) => transition.actions)
-  const transitionRaisedEvents = sortEvaluatedTransitions(machine, transitions)
-    .flatMap((transition) => transition.raisedEvents)
-
-  if (!changed) {
-    return {
-      next: stateAfterTransition,
-      actions: transitionActions,
-      raisedEvents: transitionRaisedEvents,
-      exitPaths: [],
-      entryPaths: [],
-      changed: false
-    }
-  }
-
-  const exitPaths = sortExitPaths(machine, transitions.flatMap((transition) => transition.exitPaths))
-  const entryPaths = sortEntryPaths(machine, transitions.flatMap((transition) => transition.entryPaths))
-  const exit = yield* collectStateActions<States, Events, Emits, E, R>(
-    machine,
-    state,
-    exitPaths,
-    event,
-    "exit"
-  )
-  const entry = yield* collectStateActions<States, Events, Emits, E, R>(
-    machine,
-    stateAfterTransition,
-    entryPaths,
-    event,
-    "entry"
-  )
-
-  return {
-    next: stateAfterTransition,
-    actions: [...exit.actions, ...transitionActions, ...entry.actions] as ReadonlyArray<
-      Effect.Effect<void, E, R>
-    >,
-    raisedEvents: [...exit.raisedEvents, ...transitionRaisedEvents, ...entry.raisedEvents] as ReadonlyArray<
-      Machine.EventOf<Events>
-    >,
-    exitPaths,
-    entryPaths,
-    changed: true
-  }
-})
-
-const settle: <
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema> = any,
-  const Input extends Schema.Top = typeof Schema.Void,
-  UnhandledStates extends Machine.StateIdentifier<States> = Machine.StateIdentifier<States>,
-  E = never,
-  R = never,
-  InitialE = never,
-  InitialR = never,
-  FinalStates extends Machine.StateIdentifier<States> = never,
-  Output = never
->(
-  machine: Machine<States, Events, Input, UnhandledStates, E, R, InitialE, InitialR, FinalStates, Output, Emits>,
-  state: ActiveConfiguration,
-  event: Machine.EventOf<Events>,
-  actions: Array<Effect.Effect<void, E, R>>,
-  raisedEvents: Array<Machine.EventOf<Events>>,
-  microsteps: Array<MicrostepPlan<ActiveConfiguration, Machine.EventOf<Events>, E, R>>
-) => Effect.Effect<
-  MacrostepPlan<ActiveConfiguration, Machine.EventOf<Events>, E, R, Output>,
-  E | UnhandledEventError | InfiniteTransitionError,
-  R
-> = Effect.fnUntraced(function*<
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema> = any,
-  const Input extends Schema.Top = typeof Schema.Void,
-  UnhandledStates extends Machine.StateIdentifier<States> = Machine.StateIdentifier<States>,
-  E = never,
-  R = never,
-  InitialE = never,
-  InitialR = never,
-  FinalStates extends Machine.StateIdentifier<States> = never,
-  Output = never
->(
-  machine: Machine<States, Events, Input, UnhandledStates, E, R, InitialE, InitialR, FinalStates, Output, Emits>,
-  state: ActiveConfiguration,
-  event: Machine.EventOf<Events>,
-  actions: Array<Effect.Effect<void, E, R>>,
-  raisedEvents: Array<Machine.EventOf<Events>>,
-  microsteps: Array<MicrostepPlan<ActiveConfiguration, Machine.EventOf<Events>, E, R>>
-) {
-  let currentState = state
-  let currentEvent = event
-  let shouldRunAlways = true
-  let iterations = 0
-  let raisedEventIndex = 0
-  let finalOutput: Output | undefined = undefined
-
-  while (true) {
-    const completed = completeConfiguration(machine, currentState, currentEvent)
-    currentState = completed.configuration
-    if (completed.completion !== undefined) {
-      finalOutput = completed.completion.output as Output | undefined
-      break
-    }
-
-    iterations += 1
-    if (iterations > MaxMacrostepIterations) {
-      return yield* new InfiniteTransitionError({
-        machineId: machine.id,
-        state: String(getLeafPath(machine, currentState)),
-        maxIterations: MaxMacrostepIterations
-      })
-    }
-
-    const always = shouldRunAlways
-      ? selectAlwaysTransitions<States, Events, Emits, E, R>(machine, currentState, currentEvent)
-      : []
-    if (always.length > 0) {
-      const alwaysStep: MicrostepPlan<ActiveConfiguration, Machine.EventOf<Events>, E, R> = yield* microstep(
-        machine,
-        currentState,
-        currentEvent,
-        always
-      )
-      actions.push(...alwaysStep.actions)
-      raisedEvents.push(...alwaysStep.raisedEvents)
-      microsteps.push(alwaysStep)
-      currentState = alwaysStep.next
-      shouldRunAlways = alwaysStep.changed
-      continue
-    }
-
-    const raisedEvent = raisedEvents[raisedEventIndex]
-    if (raisedEvent === undefined) {
-      break
-    }
-    raisedEventIndex += 1
-
-    currentEvent = raisedEvent
-    const raisedStep = yield* microstep(
-      machine,
-      currentState,
-      raisedEvent,
-      selectEventTransitions<States, Events, Emits, E, R>(
-        machine,
-        currentState,
-        raisedEvent as Machine.EventByTag<Events, Machine.TagOf<Events[number]>>
-      )
-    )
-    actions.push(...raisedStep.actions)
-    raisedEvents.push(...raisedStep.raisedEvents)
-    microsteps.push(raisedStep)
-    currentState = raisedStep.next
-    shouldRunAlways = true
-  }
-
-  return {
-    next: currentState,
-    actions,
-    microsteps,
-    output: finalOutput
-  }
-})
-
-const macrostep: <
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema> = any,
-  const Input extends Schema.Top = typeof Schema.Void,
-  UnhandledStates extends Machine.StateIdentifier<States> = Machine.StateIdentifier<States>,
-  E = never,
-  R = never,
-  InitialE = never,
-  InitialR = never,
-  FinalStates extends Machine.StateIdentifier<States> = never,
-  Output = never
->(
-  machine: Machine<States, Events, Input, UnhandledStates, E, R, InitialE, InitialR, FinalStates, Output, Emits>,
-  state: Machine.StateLike<States>,
-  event: Machine.EventOf<Events>
-) => Effect.Effect<
-  MacrostepPlan<Machine.Snapshot<States>, Machine.EventOf<Events>, E, R, Output>,
-  E | UnhandledEventError | InfiniteTransitionError,
-  R
-> = Effect.fnUntraced(function*<
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema> = any,
-  const Input extends Schema.Top = typeof Schema.Void,
-  UnhandledStates extends Machine.StateIdentifier<States> = Machine.StateIdentifier<States>,
-  E = never,
-  R = never,
-  InitialE = never,
-  InitialR = never,
-  FinalStates extends Machine.StateIdentifier<States> = never,
-  Output = never
->(
-  machine: Machine<States, Events, Input, UnhandledStates, E, R, InitialE, InitialR, FinalStates, Output, Emits>,
-  state: Machine.StateLike<States>,
-  event: Machine.EventOf<Events>
-) {
-  const configuration = normalizeConfiguration<States>(machine, state)
-  const snapshot = snapshotFromConfiguration<States>(machine, configuration)
-  if (isFinalState(machine, snapshot)) {
-    return {
-      next: snapshot,
-      actions: [],
-      microsteps: [],
-      output: undefined
-    }
-  }
-
-  const step = yield* microstep(
-    machine,
-    configuration,
-    event,
-    selectEventTransitions<States, Events, Emits, E, R>(
-      machine,
-      configuration,
-      event as Machine.EventByTag<Events, Machine.TagOf<Events[number]>>
-    )
-  )
-  const actions = [...step.actions]
-  const raisedEvents = [...step.raisedEvents]
-  const microsteps = [step]
-  const settled = yield* settle(machine, step.next, event, actions, raisedEvents, microsteps)
-  return {
-    next: snapshotFromConfiguration<States>(machine, settled.next),
-    actions: settled.actions,
-    microsteps: settled.microsteps.map((step) => ({
-      next: snapshotFromConfiguration<States>(machine, step.next),
-      actions: step.actions,
-      raisedEvents: step.raisedEvents,
-      exitPaths: step.exitPaths,
-      entryPaths: step.entryPaths,
-      changed: step.changed
-    })),
-    output: settled.output
-  }
-})
+): ReadonlyArray<Machine.TagOf<Events[number]>> => internalRuntime.enabled(machine, state)
 
 /**
  * Plans the next state for a state machine without running deferred actions.
@@ -3186,14 +1498,7 @@ const macrostep: <
  * @category combinators
  * @since 4.0.0
  */
-export const plan = macrostep
-
-const actionUnsafe = Effect.fnUntraced(function*<E, R>(
-  effect: Effect.Effect<void, E, R>
-) {
-  const actions = yield* DeferredActions
-  yield* actions.add(effect)
-})
+export const plan = internalRuntime.plan
 
 /**
  * Defers an effectful action until the current state machine step is planned.
@@ -3203,7 +1508,7 @@ const actionUnsafe = Effect.fnUntraced(function*<E, R>(
  */
 export const action = <E, R>(
   effect: Effect.Effect<void, E, R>
-): Effect.Effect<void, E, R> => actionUnsafe(effect) as unknown as Effect.Effect<void, E, R>
+): Effect.Effect<void, E, R> => internalRuntime.action(effect)
 
 /**
  * Returns the typed runtime capability for the current state machine.
@@ -3215,12 +1520,7 @@ export const runtime = <const Protocol extends Runtime.Protocol = {}>(): Effect.
   Runtime<Runtime.Events<Protocol>, Runtime.Emits<Protocol>>,
   never,
   Runtime.Requirement<Runtime.Events<Protocol>, Runtime.Emits<Protocol>>
-> =>
-  RuntimeContext as unknown as Effect.Effect<
-    Runtime<Runtime.Events<Protocol>, Runtime.Emits<Protocol>>,
-    never,
-    Runtime.Requirement<Runtime.Events<Protocol>, Runtime.Emits<Protocol>>
-  >
+> => internalRuntime.runtime<Protocol>()
 
 /**
  * Returns the current actor runtime scope when a state machine is running as
@@ -3233,7 +1533,7 @@ export const actorRuntime = <Event = unknown>(): Effect.Effect<
   ActorModule.ActorScope<Event>,
   never,
   ActorRuntime
-> => Effect.map(ActorRuntime, (runtime) => runtime as ActorModule.ActorScope<Event>)
+> => internalRuntime.actorRuntime<Event>()
 
 /**
  * Creates a typed parent-local address for a child actor.
@@ -3384,264 +1684,7 @@ export const toActorLogic: <
   ExcludeCompatibleRuntime<Exclude<InitialR | R, ActorRuntime>, Machine.EventOf<Events>, Machine.EmitOf<Emits>>,
   Output | undefined,
   InitialE | StartupError
-> = <
-  const States extends Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.TaggedSchema> = any,
-  const Input extends Schema.Top = typeof Schema.Void,
-  UnhandledStates extends Machine.StateIdentifier<States> = Machine.StateIdentifier<States>,
-  E = never,
-  R = never,
-  InitialE = never,
-  InitialR = never,
-  FinalStates extends Machine.StateIdentifier<States> = never,
-  Output = never
->(
-  machine: Machine<States, Events, Input, UnhandledStates, E, R, InitialE, InitialR, FinalStates, Output, Emits>,
-  ...args: [...Machine.InputArgs<Input>]
-) =>
-  ActorModule.make({
-    initial: (scope) =>
-      provideActorRuntime(
-        Effect.gen(function*() {
-          const planned = yield* planInitial(machine, ...args)
-          yield* runActions(planned.actions, makeActorRuntime<Machine.EventOf<Events>, Machine.EmitOf<Emits>>(scope))
-          return planned.state
-        }),
-        scope
-      ),
-    run: (context) =>
-      provideActorRuntime(
-        Effect.gen(function*() {
-          const { receive, state, setState } = context
-          let done = false
-          let output: Output | undefined = undefined
-
-          const initialState = yield* state
-          if (isFinalState(machine, initialState)) {
-            return getFinalOutput<States, Events, Output>(
-              machine,
-              initialState,
-              InitialEvent as Machine.EventOf<Events>
-            )
-          }
-
-          const invokeSessions = yield* Ref.make<HashMap.HashMap<string, InvokeSession>>(HashMap.empty())
-          const makeInvokeSessionKey = (path: string, id: string): string => `${path.length}:${path}${id}`
-          const makeInvokeChildId = (path: string, id: string): string =>
-            `StateMachine.invoke:${makeInvokeSessionKey(path, id)}`
-          const isCurrentInvoke = (key: string, token: symbol): Effect.Effect<boolean> =>
-            Ref.get(invokeSessions).pipe(
-              Effect.map((sessions) => {
-                const current = HashMap.get(sessions, key)
-                return Option.isSome(current) && current.value.token === token
-              })
-            )
-          const stopInvoke = (key: string, exit: Exit.Exit<unknown, unknown>): Effect.Effect<void> =>
-            Ref.modify(invokeSessions, (sessions) => {
-              const current = HashMap.get(sessions, key)
-              return Option.isSome(current)
-                ? [current.value, HashMap.remove(sessions, key)] as const
-                : [undefined, sessions] as const
-            }).pipe(
-              Effect.flatMap((session) =>
-                session === undefined
-                  ? Effect.void
-                  : Scope.close(session.scope, exit).pipe(
-                    Effect.andThen(context.stopChild(session.childId))
-                  )
-              )
-            )
-          const clearInvoke = (key: string, token: symbol, exit: Exit.Exit<unknown, unknown>): Effect.Effect<void> =>
-            Ref.modify(invokeSessions, (sessions) => {
-              const current = HashMap.get(sessions, key)
-              return Option.isSome(current) && current.value.token === token
-                ? [current.value, HashMap.remove(sessions, key)] as const
-                : [undefined, sessions] as const
-            }).pipe(
-              Effect.flatMap((session) => session === undefined ? Effect.void : Scope.close(session.scope, exit))
-            )
-          const stopAllInvokes = (exit: Exit.Exit<unknown, unknown>): Effect.Effect<void> =>
-            Ref.modify(invokeSessions, (sessions) => [HashMap.toEntries(sessions), HashMap.empty()] as const).pipe(
-              Effect.flatMap((sessions) =>
-                Effect.all(
-                  sessions.map(([, session]) =>
-                    Scope.close(session.scope, exit).pipe(
-                      Effect.andThen(context.stopChild(session.childId))
-                    )
-                  ),
-                  { discard: true, concurrency: "unbounded" }
-                )
-              )
-            )
-          const startInvokeWatchers = (
-            config: AnyInvokeConfig,
-            child: ActorModule.Actor<any, any, any, any>,
-            key: string,
-            token: symbol,
-            scope: Scope.Closeable
-          ): Effect.Effect<void> =>
-            Effect.gen(function*() {
-              if (config.snapshot !== undefined) {
-                const mapSnapshot = config.snapshot
-                yield* child.changes.pipe(
-                  Stream.filter((snapshot) => snapshot.status === "active"),
-                  Stream.runForEach((snapshot) =>
-                    isCurrentInvoke(key, token).pipe(
-                      Effect.flatMap((isCurrent) => {
-                        if (!isCurrent) {
-                          return Effect.void
-                        }
-                        const mappedEvent = mapSnapshot({ id: config.id, snapshot })
-                        return mappedEvent === undefined
-                          ? Effect.void
-                          : context.self.send(mappedEvent as Machine.EventOf<Events>)
-                      })
-                    )
-                  ),
-                  Effect.forkIn(scope),
-                  Effect.asVoid
-                )
-              }
-              if (config.event !== undefined) {
-                const mapEvent = config.event
-                yield* ActorModule.watch(child).pipe(
-                  Stream.runForEach((outcome) =>
-                    isCurrentInvoke(key, token).pipe(
-                      Effect.flatMap((isCurrent) => {
-                        if (!isCurrent) {
-                          return Effect.void
-                        }
-                        const mappedEvent = mapEvent({ id: config.id, outcome })
-                        return mappedEvent === undefined
-                          ? Effect.void
-                          : context.self.send(mappedEvent as Machine.EventOf<Events>)
-                      })
-                    )
-                  ),
-                  Effect.forkIn(scope),
-                  Effect.asVoid
-                )
-              }
-            })
-          const startInvoke = <StateId extends Machine.StateIdentifier<States>>(
-            path: StateId,
-            config: AnyInvokeConfig,
-            state: Machine.StateByIdentifier<States, StateId>,
-            event: Machine.EventOf<Events>
-          ) =>
-            Effect.gen(function*() {
-              const token = Symbol()
-              const invokeId = String(config.id)
-              const key = makeInvokeSessionKey(path, invokeId)
-              const childId = makeInvokeChildId(path, invokeId)
-              const child = yield* context.spawn(
-                config.src({
-                  state,
-                  event,
-                  runtime: runtimeFor<Machine.EventOf<Events>, Machine.EmitOf<Emits>>()
-                }),
-                { id: childId }
-              ).pipe(
-                Effect.onExit((exit) =>
-                  Exit.isFailure(exit) ? clearInvoke(key, token, Exit.failCause(exit.cause)) : Effect.void
-                )
-              )
-              const scope = yield* Scope.make("parallel")
-              yield* Ref.update(invokeSessions, (sessions) => HashMap.set(sessions, key, { token, scope, childId }))
-              yield* startInvokeWatchers(config, child, key, token, scope)
-            })
-          const startInvokes = (
-            state: Machine.Snapshot<States>,
-            paths: ReadonlyArray<string>,
-            event: Machine.EventOf<Events>
-          ): Effect.Effect<void, E, R> => {
-            const configuration = normalizeConfiguration(machine, state)
-            return Effect.all(
-              sortEntryPaths(machine, paths)
-                .filter((path) => configuration.active.has(path))
-                .flatMap((path) =>
-                  getInvokes(getStateConfigByPath(machine, path)).map((config) =>
-                    startInvoke(
-                      path as Machine.StateIdentifier<States>,
-                      config,
-                      getActiveValue(configuration, path) as Machine.StateByIdentifier<
-                        States,
-                        Machine.StateIdentifier<States>
-                      >,
-                      event
-                    ) as Effect.Effect<void, E, R>
-                  )
-                ),
-              { discard: true }
-            )
-          }
-          const stopInvokes = (paths: ReadonlyArray<string>): Effect.Effect<void> =>
-            Effect.all(
-              sortExitPaths(machine, paths).flatMap((path) =>
-                getInvokes(getStateConfigByPath(machine, path)).map((config) =>
-                  stopInvoke(makeInvokeSessionKey(path, String(config.id)), Exit.void)
-                )
-              ),
-              { discard: true, concurrency: "unbounded" }
-            )
-
-          return yield* Effect.gen(function*() {
-            yield* startInvokes(
-              initialState,
-              getInitialEntryPaths(machine, normalizeConfiguration(machine, initialState)),
-              InitialEvent as Machine.EventOf<Events>
-            )
-
-            yield* Effect.whileLoop({
-              while: () => !done,
-              body: () =>
-                Effect.gen(function*() {
-                  const event = yield* receive
-                  const current = yield* state
-                  const planned = yield* macrostep(machine, current, event)
-                  const changed = planned.microsteps.some((step) => step.changed)
-                  const exitPaths = planned.microsteps.flatMap((step) => step.exitPaths)
-                  const entryPaths = planned.microsteps.flatMap((step) => step.entryPaths)
-
-                  if (changed) {
-                    yield* stopInvokes(exitPaths)
-                  }
-                  yield* setState(planned.next)
-                  yield* runActions(
-                    planned.actions,
-                    makeActorRuntime<Machine.EventOf<Events>, Machine.EmitOf<Emits>>(context)
-                  )
-
-                  if (isFinalState(machine, planned.next)) {
-                    done = true
-                    output = planned.output
-                    yield* stopAllInvokes(Exit.succeed(output))
-                  } else {
-                    if (changed) {
-                      yield* startInvokes(planned.next, entryPaths, event)
-                    }
-                    yield* Effect.yieldNow
-                  }
-                }),
-              step: () => undefined
-            })
-
-            return output
-          }).pipe(
-            Effect.onExit((exit) => stopAllInvokes(exit))
-          )
-        }),
-        context
-      )
-  }) as ActorModule.ActorLogic<
-    Machine.Snapshot<States>,
-    Machine.EventOf<Events>,
-    E | UnhandledEventError | InfiniteTransitionError,
-    ExcludeCompatibleRuntime<Exclude<InitialR | R, ActorRuntime>, Machine.EventOf<Events>, Machine.EmitOf<Emits>>,
-    Output | undefined,
-    InitialE | StartupError
-  >
+> = internalActor.toActorLogic as any
 
 /**
  * Starts a state machine as an actor.
@@ -3687,7 +1730,4 @@ export const start: <
   >,
   InitialE | StartupError,
   ExcludeCompatibleRuntime<Exclude<InitialR | R, ActorRuntime>, Machine.EventOf<Events>, Machine.EmitOf<Emits>>
-> = (
-  machine,
-  ...args
-) => ActorModule.start(toActorLogic(machine, ...args))
+> = internalActor.start as any
